% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_matching_functions.R
\name{Match_prop_ICD}
\alias{Match_prop_ICD}
\title{Function performing propensity score based matching using functions from the
MatchIt package.}
\usage{
Match_prop_ICD(
  treated_group,
  possible_controls,
  age_dif = 10,
  BMI_dif = 5,
  model = "mod_0",
  caliper = 0.2
)
}
\arguments{
\item{treated_group}{a dataframe containing the data for the exposed group,
the dataframe should at least contain a column called "age_admission" with
the age of the individual at admission into the hospital, "BMI" with the BMI,
"gender" with gender of the individual, "sentence" with a string with all
ICD-10 codes in sequantial order, and a column for each individual ICD-10
code where the columns are ordered in sequential order. For running the propensity
score models that include ICD-10 codes, these information needs to be provided in
the columns as well (see example for how to create these data).}

\item{possible_controls}{A dataframe of potential controls with the same structure
as `treated_group`.}

\item{age_dif}{Maximum permissible age difference for matching (default is 10 years).}

\item{BMI_dif}{Maximum permissible BMI difference for matching (default is 5 points).}

\item{model}{A string identifier for the model used to calculate propensity scores.
Different models use different sets of covariates for more refined control.}
}
\value{
A list containing hospitalizations in the matched cohort, the hospitalizations
for whom a match was found, the matchit output object for further analysis,
the dataframe with added propensity scores and inverse probability
weights (IPW) adjusted for case status, and the number of individuals that were matched.
}
\description{
Function performing propensity score based matching using functions from the
MatchIt package.
}
