% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_cox_hazard.R
\name{perform_mortality_analysis}
\alias{perform_mortality_analysis}
\title{Function to perform cox proportional hazard analysis given the output of one
of the matching functions}
\usage{
perform_mortality_analysis(matched_object, figure = TRUE)
}
\arguments{
\item{matched_object}{result of WEMatch or Match_prop_ICD}

\item{figure}{should a figure be produced? (default is TRUE)}
}
\description{
Function to perform cox proportional hazard analysis given the output of one
of the matching functions
}
