% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training_data}
\alias{training_data}
\title{Training data for w2v model}
\format{
## `training ICD-10 data`
A data frame with 20,000 rows and 1 column:
\describe{
  \item{sentence}{All ICD-10 codes combined per hospitalisation}
}
}
\source{
Simulated
}
\usage{
training_data
}
\description{
These are simulated data where the ICD-10 codes and additional variables are all
simulated independent of each other. These data do not have any clinical value.
}
\keyword{datasets}
