% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_matching_functions.R
\name{WEMatch}
\alias{WEMatch}
\title{Function performing the WEMatch algorithm.}
\usage{
WEMatch(exposed_group, possible_controls, W2V_model, age_dif = 10, BMI_dif = 5)
}
\arguments{
\item{exposed_group}{a dataframe containing the data for the exposed group,
the dataframe should at least contain a column called "age_admission" with
the age of the individual at admission into the hospital, "BMI" with the BMI,
"gender" with gender of the individual, "sentence" with a string with all
ICD-10 codes in sequantial order, and a column for each individual ICD-10
code where the columns are ordered in sequential order.}

\item{possible_controls}{A dataframe of potential controls with the same structure
as `exposed_group`.}

\item{W2V_model}{A trained Word2Vec model object.}

\item{age_dif}{Maximum allowed difference in age for matching (default is 10).}

\item{BMI_dif}{Maximum allowed difference in BMI for matching (default is 5).}
}
\value{
A list containing hospitalizations in the matched cohort, the hospitalizations
for whom a match was found, the number of perfect matches, the number of matches based
on sequential ICD-10 order, and a vector with for each match the number of ICD-10 codes
which match perfectly.
}
\description{
Function performing the WEMatch algorithm.
}
