% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_matching_functions.R
\name{Match_prop_ICD}
\alias{Match_prop_ICD}
\title{Title}
\usage{
Match_prop_ICD(
  exposed_group,
  possible_controls,
  method = "nearest",
  age_dif = 10,
  BMI_dif = 5,
  model = "mod_0"
)
}
\arguments{
\item{exposed_group}{a dataframe containing the data for the exposed group,
the dataframe should at least contain a column called "age_admission" with
the age of the individual at admission into the hospital, "BMI" with the BMI,
"gender" with gender of the individual, "sentence" with a string with all
ICD-10 codes in sequantial order, and a column for each individual ICD-10
code where the columns are ordered in sequential order.}

\item{possible_controls}{A dataframe of potential controls with the same structure
as `exposed_group`.}

\item{method}{A string specifying the distance method for finding matches; default is 'nearest'.
Options include 'nearest' for nearest neighbor matching and 'linear_nearest' for linear
nearest matching which uses logit transformations.}

\item{age_dif}{Maximum permissible age difference for matching (default is 10 years).}

\item{BMI_dif}{Maximum permissible BMI difference for matching (default is 5 points).}

\item{model}{A string identifier for the model used to calculate propensity scores.
Different models use different sets of covariates for more refined control.}
}
\value{
A list containing hospitalizations in the matched cohort, the hospitalizations
for whom a match was found, the matchit output object for further analysis,
and the dataframe with added propensity scores and inverse probability
weights (IPW) adjusted for case status.
}
\description{
Title
}
